# gateway/event.swagger.json
{
    "openapi": "3.0.0",
    "paths": {
...
      "/events": {
        "post": {
          "operationId": "EventsController_create",
          "parameters": [],
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateEventDto"
                }
              }
            }
          },
          "responses": {
            "201": {
              "description": "이벤트가 성공적으로 생성되었습니다.",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/EventResponseDto"
                  }
                }
              }
            },
            "400": {
              "description": "유효하지 않은 조건 ID 또는 보상 ID입니다."
            },
            "409": {
              "description": "이미 존재하는 이벤트 ID입니다."
            }
          },
          "summary": "새 이벤트 생성",
          "tags": [
            "events"
          ]
        },
        "get": {
          "operationId": "EventsController_findAll",
          "parameters": [],
          "responses": {
            "200": {
              "description": "모든 이벤트 목록을 반환합니다.",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/EventResponseDto"
                    }
                  }
                }
              }
            }
          },
          "summary": "모든 이벤트 조회",
          "tags": [
            "events"
          ]
        }
      },
      "/events/{eventId}": {
        "get": {
          "operationId": "EventsController_findOne",
          "parameters": [
            {
              "name": "eventId",
              "required": true,
              "in": "path",
              "schema": {
                "type": "string"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "특정 이벤트 정보를 반환합니다.",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/EventResponseDto"
                  }
                }
              }
            },
            "404": {
              "description": "이벤트를 찾을 수 없습니다."
            }
          },
          "summary": "특정 이벤트 조회",
          "tags": [
            "events"
          ]
        },
        "put": {
          "operationId": "EventsController_update",
          "parameters": [
            {
              "name": "eventId",
              "required": true,
              "in": "path",
              "schema": {
                "type": "string"
              }
            }
          ],
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateEventDto"
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "이벤트가 성공적으로 업데이트되었습니다.",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/EventResponseDto"
                  }
                }
              }
            },
            "400": {
              "description": "유효하지 않은 조건 ID 또는 보상 ID입니다."
            },
            "404": {
              "description": "이벤트를 찾을 수 없습니다."
            }
          },
          "summary": "이벤트 전체 업데이트",
          "tags": [
            "events"
          ]
        },
        "patch": {
          "operationId": "EventsController_partialUpdate",
          "parameters": [
            {
              "name": "eventId",
              "required": true,
              "in": "path",
              "schema": {
                "type": "string"
              }
            }
          ],
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateEventDto"
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "이벤트가 성공적으로 부분 업데이트되었습니다.",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/EventResponseDto"
                  }
                }
              }
            },
            "400": {
              "description": "유효하지 않은 조건 ID 또는 보상 ID입니다."
            },
            "404": {
              "description": "이벤트를 찾을 수 없습니다."
            }
          },
          "summary": "이벤트 부분 업데이트",
          "tags": [
            "events"
          ]
        },
        "delete": {
          "operationId": "EventsController_remove",
          "parameters": [
            {
              "name": "eventId",
              "required": true,
              "in": "path",
              "schema": {
                "type": "string"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "이벤트가 성공적으로 삭제되었습니다."
            },
            "404": {
              "description": "이벤트를 찾을 수 없습니다."
            }
          },
          "summary": "이벤트 삭제",
          "tags": [
            "events"
          ]
        }
      },
      "/events/{eventId}/activate": {
        "patch": {
          "operationId": "EventsController_activate",
          "parameters": [
            {
              "name": "eventId",
              "required": true,
              "in": "path",
              "schema": {
                "type": "string"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "이벤트가 성공적으로 활성화되었습니다.",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/EventResponseDto"
                  }
                }
              }
            },
            "404": {
              "description": "이벤트를 찾을 수 없습니다."
            }
          },
          "summary": "이벤트 활성화",
          "tags": [
            "events"
          ]
        }
      },
      "/events/{eventId}/deactivate": {
        "patch": {
          "operationId": "EventsController_deactivate",
          "parameters": [
            {
              "name": "eventId",
              "required": true,
              "in": "path",
              "schema": {
                "type": "string"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "이벤트가 성공적으로 비활성화되었습니다.",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/EventResponseDto"
                  }
                }
              }
            },
            "404": {
              "description": "이벤트를 찾을 수 없습니다."
            }
          },
          "summary": "이벤트 비활성화",
          "tags": [
            "events"
          ]
        }
      },
...
      "/reward-history/user/{userId}": {
        "get": {
          "operationId": "RewardHistoryController_findByUserId",
          "parameters": [
            {
              "name": "userId",
              "required": true,
              "in": "path",
              "description": "사용자 ID",
              "schema": {
                "type": "string"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "사용자별 보상 이력 목록",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/RewardHistoryResponseDto"
                    }
                  }
                }
              }
            }
          },
          "summary": "사용자별 보상 이력 조회",
          "tags": [
            "reward-history"
          ]
        }
      },
...
      "/user-event-rewards/user/{userId}": {
        "get": {
          "operationId": "UserEventRewardsController_findAllByUserId",
          "parameters": [
            {
              "name": "userId",
              "required": true,
              "in": "path",
              "description": "유저 ID",
              "schema": {
                "type": "string"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "유저의 모든 이벤트 보상 요청 목록을 반환합니다.",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/UserEventRewardRequestResponseDto"
                    }
                  }
                }
              }
            }
          },
          "summary": "유저의 모든 이벤트 보상 요청 조회",
          "tags": [
            "user-event-rewards"
          ]
        }
      },
      "/user-event-rewards/user/{userId}/event/{eventId}": {
        "get": {
          "operationId": "UserEventRewardsController_findAllByUserIdAndEventId",
          "parameters": [
            {
              "name": "userId",
              "required": true,
              "in": "path",
              "description": "유저 ID",
              "schema": {
                "type": "string"
              }
            },
            {
              "name": "eventId",
              "required": true,
              "in": "path",
              "description": "이벤트 ID",
              "schema": {
                "type": "string"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "특정 이벤트에 대한 유저의 보상 요청 목록을 반환합니다.",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/UserEventRewardRequestResponseDto"
                    }
                  }
                }
              }
            }
          },
          "summary": "특정 이벤트에 대한 유저의 보상 요청 조회",
          "tags": [
            "user-event-rewards"
          ]
        }
      },
...
    },
    "info": {
      "title": "이벤트 서비스 API",
      "description": "이벤트 관리 시스템을 위한 API 문서",
      "version": "1.0",
      "contact": {
  
      }
    },
...
    "servers": [],
    "components": {
      "securitySchemes": {
        "JWT": {
          "scheme": "bearer",
          "bearerFormat": "JWT",
          "type": "http",
          "name": "JWT",
          "description": "JWT 토큰을 입력하세요",
          "in": "header"
        }
      },
      "schemas": {
        "CreateEventDto": {
          "type": "object",
          "properties": {
            "eventId": {
              "type": "string",
              "description": "이벤트 ID",
              "example": "event-001"
            },
            "title": {
              "type": "string",
              "description": "이벤트 제목",
              "example": "7일 출석 이벤트"
            },
            "category": {
              "type": "string",
              "description": "이벤트 카테고리",
              "enum": [
                "attendance",
                "social",
                "economy",
                "mission",
                "special"
              ],
              "example": "attendance"
            },
            "conditionIds": {
              "description": "이벤트 조건 ID 목록",
              "example": [
                "60d21b4667d0d8992e610c85",
                "60d21b4667d0d8992e610c86"
              ],
              "type": "array",
              "items": {
                "type": "string"
              }
            },
            "rewardIds": {
              "description": "이벤트 보상 ID 목록",
              "example": [
                "60d21b4667d0d8992e610c87",
                "60d21b4667d0d8992e610c88"
              ],
              "type": "array",
              "items": {
                "type": "string"
              }
            },
            "trigger": {
              "type": "string",
              "description": "이벤트 트리거 유형",
              "enum": [
                "manual",
                "auto"
              ],
              "default": "manual"
            },
            "deliveryType": {
              "type": "string",
              "description": "보상 전달 유형",
              "enum": [
                "immediate",
                "manual_claim",
                "scheduled"
              ],
              "default": "manual_claim"
            },
            "startAt": {
              "format": "date-time",
              "type": "string",
              "description": "이벤트 시작 시간",
              "example": "2025-01-01T00:00:00Z"
            },
            "endAt": {
              "format": "date-time",
              "type": "string",
              "description": "이벤트 종료 시간",
              "example": "2025-01-31T23:59:59Z"
            },
            "isActive": {
              "type": "boolean",
              "description": "이벤트 활성화 여부",
              "default": true
            }
          },
          "required": [
            "eventId",
            "title",
            "category",
            "conditionIds",
            "rewardIds",
            "trigger",
            "deliveryType",
            "startAt",
            "endAt",
            "isActive"
          ]
        },
        "ConditionResponseDto": {
          "type": "object",
          "properties": {
            "id": {
              "type": "string",
              "description": "조건 ID",
              "example": "60a12d5c8f45e32b4c9cde1f"
            },
            "category": {
              "type": "string",
              "description": "조건 카테고리",
              "enum": [
                "attendance",
                "social",
                "economy",
                "mission",
                "special"
              ],
              "example": "attendance"
            },
            "subType": {
              "type": "string",
              "description": "조건 하위 타입",
              "enum": [
                "total_days",
                "after_signup",
                "invite_success",
                "meso_spent",
                "cash_topup",
                "item_collect",
                "quest_clear",
                "return_after_days"
              ],
              "example": "total_days"
            },
            "target": {
              "type": "object",
              "description": "조건 기준값 (예: 3일, 아이템 ID 등)",
              "example": 3
            },
            "description": {
              "type": "string",
              "description": "조건 설명",
              "example": "3일 연속 출석 시 보상 지급"
            },
            "createdAt": {
              "format": "date-time",
              "type": "string",
              "description": "생성 시간",
              "example": "2025-05-20T00:00:00.000Z"
            },
            "updatedAt": {
              "format": "date-time",
              "type": "string",
              "description": "수정 시간",
              "example": "2025-05-20T00:00:00.000Z"
            }
          },
          "required": [
            "id",
            "category",
            "subType",
            "createdAt",
            "updatedAt"
          ]
        },
        "RewardResponseDto": {
          "type": "object",
          "properties": {
            "id": {
              "type": "string",
              "description": "보상 ID",
              "example": "60d21b4667d0d8992e610c85"
            },
            "type": {
              "type": "string",
              "description": "보상 타입",
              "enum": [
                "item",
                "meso",
                "mPoint"
              ],
              "example": "item"
            },
            "quantity": {
              "type": "number",
              "description": "보상 수량",
              "example": 1
            },
            "itemId": {
              "type": "string",
              "description": "아이템 ID (타입이 아이템인 경우에만 존재)",
              "example": "60d21b4667d0d8992e610c85"
            },
            "description": {
              "type": "string",
              "description": "보상 설명",
              "example": "레어 아이템 보상"
            },
            "createdAt": {
              "format": "date-time",
              "type": "string",
              "description": "생성 일시",
              "example": "2025-05-20T00:37:58.000Z"
            },
            "updatedAt": {
              "format": "date-time",
              "type": "string",
              "description": "수정 일시",
              "example": "2025-05-20T00:37:58.000Z"
            }
          },
          "required": [
            "id",
            "type",
            "quantity",
            "createdAt",
            "updatedAt"
          ]
        },
        "EventResponseDto": {
          "type": "object",
          "properties": {
            "eventId": {
              "type": "string",
              "description": "이벤트 ID",
              "example": "event-001"
            },
            "title": {
              "type": "string",
              "description": "이벤트 제목",
              "example": "7일 출석 이벤트"
            },
            "category": {
              "type": "string",
              "description": "이벤트 카테고리",
              "enum": [
                "attendance",
                "social",
                "economy",
                "mission",
                "special"
              ],
              "example": "attendance"
            },
            "conditions": {
              "description": "이벤트 조건 목록",
              "type": "array",
              "items": {
                "$ref": "#/components/schemas/ConditionResponseDto"
              }
            },
            "rewards": {
              "description": "이벤트 보상 목록",
              "type": "array",
              "items": {
                "$ref": "#/components/schemas/RewardResponseDto"
              }
            },
            "trigger": {
              "type": "string",
              "description": "트리거 타입",
              "enum": [
                "manual",
                "auto"
              ],
              "example": "manual"
            },
            "deliveryType": {
              "type": "string",
              "description": "보상 전달 타입",
              "enum": [
                "immediate",
                "manual_claim",
                "scheduled"
              ],
              "example": "manual_claim"
            },
            "startAt": {
              "format": "date-time",
              "type": "string",
              "description": "이벤트 시작 시간",
              "example": "2025-01-01T00:00:00Z"
            },
            "endAt": {
              "format": "date-time",
              "type": "string",
              "description": "이벤트 종료 시간",
              "example": "2025-01-31T23:59:59Z"
            },
            "isActive": {
              "type": "boolean",
              "description": "이벤트 활성화 여부",
              "example": true
            },
            "createdAt": {
              "format": "date-time",
              "type": "string",
              "description": "생성 시간",
              "example": "2025-05-20T00:00:00.000Z"
            },
            "updatedAt": {
              "format": "date-time",
              "type": "string",
              "description": "수정 시간",
              "example": "2025-05-20T00:00:00.000Z"
            }
          },
          "required": [
            "eventId",
            "title",
            "category",
            "conditions",
            "rewards",
            "trigger",
            "deliveryType",
            "startAt",
            "endAt",
            "isActive",
            "createdAt",
            "updatedAt"
          ]
        },
        "UpdateEventDto": {
          "type": "object",
          "properties": {
            "title": {
              "type": "string",
              "description": "이벤트 제목",
              "example": "7일 출석 이벤트"
            },
            "category": {
              "type": "string",
              "description": "이벤트 카테고리",
              "enum": [
                "attendance",
                "social",
                "economy",
                "mission",
                "special"
              ],
              "example": "attendance"
            },
            "conditionIds": {
              "description": "이벤트 조건 ID 목록",
              "example": [
                "60d21b4667d0d8992e610c85",
                "60d21b4667d0d8992e610c86"
              ],
              "type": "array",
              "items": {
                "type": "string"
              }
            },
            "rewardIds": {
              "description": "이벤트 보상 ID 목록",
              "example": [
                "60d21b4667d0d8992e610c87",
                "60d21b4667d0d8992e610c88"
              ],
              "type": "array",
              "items": {
                "type": "string"
              }
            },
            "trigger": {
              "type": "string",
              "description": "이벤트 트리거 유형",
              "enum": [
                "manual",
                "auto"
              ],
              "default": "manual"
            },
            "deliveryType": {
              "type": "string",
              "description": "보상 전달 유형",
              "enum": [
                "immediate",
                "manual_claim",
                "scheduled"
              ],
              "default": "manual_claim"
            },
            "startAt": {
              "format": "date-time",
              "type": "string",
              "description": "이벤트 시작 시간",
              "example": "2025-01-01T00:00:00Z"
            },
            "endAt": {
              "format": "date-time",
              "type": "string",
              "description": "이벤트 종료 시간",
              "example": "2025-01-31T23:59:59Z"
            },
            "isActive": {
              "type": "boolean",
              "description": "이벤트 활성화 여부",
              "default": true
            }
          }
        },
...
        "RewardHistoryResponseDto": {
          "type": "object",
          "properties": {
            "id": {
              "type": "string",
              "description": "보상 이력 ID",
              "example": "60d21b4667d0d8992e610c85"
            },
            "userId": {
              "type": "string",
              "description": "사용자 ID",
              "example": "user123"
            },
            "eventId": {
              "type": "string",
              "description": "이벤트 ID",
              "example": "60d21b4667d0d8992e610c85"
            },
            "reward": {
              "description": "보상 목록",
              "type": "array",
              "items": {
                "$ref": "#/components/schemas/RewardResponseDto"
              }
            },
            "claimedAt": {
              "format": "date-time",
              "type": "string",
              "description": "수령 일시",
              "example": "2025-05-20T00:37:58.000Z"
            },
            "deliveryType": {
              "type": "string",
              "description": "지급 방식",
              "enum": [
                "immediate",
                "manual_claim",
                "scheduled"
              ],
              "example": "manual_claim"
            },
            "status": {
              "type": "string",
              "description": "보상 상태",
              "enum": [
                "pending",
                "success",
                "failed"
              ],
              "example": "success"
            },
            "reason": {
              "type": "string",
              "description": "실패 이유 (실패 시에만 존재)",
              "example": "이미 수령한 보상입니다"
            },
            "createdAt": {
              "format": "date-time",
              "type": "string",
              "description": "생성 일시",
              "example": "2025-05-20T00:37:58.000Z"
            },
            "updatedAt": {
              "format": "date-time",
              "type": "string",
              "description": "수정 일시",
              "example": "2025-05-20T00:37:58.000Z"
            }
          },
          "required": [
            "id",
            "userId",
            "eventId",
            "reward",
            "claimedAt",
            "deliveryType",
            "status",
            "createdAt",
            "updatedAt"
          ]
        },
...
        "UserEventRewardRequestResponseDto": {
          "type": "object",
          "properties": {
            "id": {
              "type": "string",
              "description": "요청 ID",
              "example": "60d21b4667d0d8992e610c90"
            },
            "userId": {
              "type": "string",
              "description": "유저 ID",
              "example": "user-123"
            },
            "eventId": {
              "type": "string",
              "description": "이벤트 ID",
              "example": "event-001"
            },
            "rewardId": {
              "type": "string",
              "description": "보상 ID",
              "example": "60d21b4667d0d8992e610c87"
            },
            "trigger": {
              "type": "string",
              "description": "트리거 타입",
              "enum": [
                "manual",
                "auto"
              ],
              "example": "manual"
            },
            "deliveryType": {
              "type": "string",
              "description": "보상 전달 타입",
              "enum": [
                "immediate",
                "manual_claim",
                "scheduled"
              ],
              "example": "manual_claim"
            },
            "rewardSnapshot": {
              "type": "object",
              "description": "보상 스냅샷 정보",
              "example": {
                "type": "item",
                "quantity": 1,
                "itemId": "item-001",
                "description": "레어 아이템"
              }
            },
            "status": {
              "type": "string",
              "description": "보상 상태",
              "enum": [
                "pending",
                "success",
                "failed"
              ],
              "example": "pending"
            },
            "requestedAt": {
              "format": "date-time",
              "type": "string",
              "description": "유저가 보상을 요청한 시간",
              "example": "2025-01-01T12:00:00Z"
            },
            "reason": {
              "type": "string",
              "description": "보상 처리 결과에 대한 이유 (실패 시 오류 메시지 등)",
              "example": "이미 보상을 받았습니다."
            },
            "createdAt": {
              "format": "date-time",
              "type": "string",
              "description": "생성 시간",
              "example": "2025-01-01T12:00:00Z"
            },
            "updatedAt": {
              "format": "date-time",
              "type": "string",
              "description": "수정 시간",
              "example": "2025-01-01T12:00:00Z"
            }
          },
          "required": [
            "id",
            "userId",
            "eventId",
            "rewardId",
            "trigger",
            "deliveryType",
            "rewardSnapshot",
            "status",
            "requestedAt",
            "reason",
            "createdAt",
            "updatedAt"
          ]
        }
      }
    }
  }